+++ portlets/knowledge-base-portlet/docroot/WEB-INF/service/com/liferay/knowledgebase/model/ArticleSoap.java
@@ -224,11 +224,11 @@ public class ArticleSoap implements Serializable {
 		_description = description;
 	}
 
-	public long getPriority() {
+	public double getPriority() {
 		return _priority;
 	}
 
-	public void setPriority(long priority) {
+	public void setPriority(double priority) {
 		_priority = priority;
 	}
 
@@ -287,7 +287,7 @@ public class ArticleSoap implements Serializable {
 	private String _title;
 	private String _content;
 	private String _description;
-	private long _priority;
+	private double _priority;
 	private int _latest;
 	private int _status;
 	private long _statusByUserId;
